Add sync implementation
