+++ Cargo.toml
@@ -2,7 +2,7 @@
 name = "lis2dh"
 version = "0.1.0"
 edition = "2024"
-description = "Driver crate for ST LIS2DH12 accelerometer"
+description = "Driver crate for the ST LIS2DH12 accelerometer"
 repository = "https://github.com/andresovela/lis2dh-rs"
 documentation = "https://docs.rs/lis2dh"
 readme = "README.md"
@@ -13,4 +13,7 @@ exclude = [".github"]
 
 [dependencies]
 embedded-hal = "1.0.0"
-embedded-hal-async = "1.0.0"
+embedded-hal-async = { version = "1.0.0", optional = true }
+
+[features]
+async = ["dep:embedded-hal-async"]
